Created Student to replace Grade and added grade updating
Rather than have a Grade entity to keep track of what grade a User is
in for a given Year, a District will have many Rosters and a Roster
will have a Year associated with it along with many Students. A Student
has a numeric grade and a User associated with it.

The District roster editor now allows students whom have been added to
have a grade set for them on an individual basis. The grades are then
persisted.